updating and moving the DB logic in separate package
