+++ src/main/java/com/musala/simple/students/db/Main.java
@@ -1,14 +1,17 @@
-package com.musala.simple.students;
+package com.musala.simple.students.db;
 
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
-import java.io.FileWriter;
 import java.io.IOException;
 
 import com.google.gson.Gson;
 import com.google.gson.JsonSyntaxException;
+import com.mongodb.DB;
+import com.mongodb.DBCollection;
+import com.mongodb.DBCursor;
+import com.mongodb.MongoClient;
 
 public class Main {
 
@@ -16,13 +19,14 @@ public class Main {
 
 		Gson gson = new Gson();
 		BufferedReader br = null;
-		ClassLoader classLoader = new Main().getClass().getClassLoader();
-		File bkpFile = new File(classLoader.getResource("backup.json").getFile());
+		MongoClient mongoClient = new MongoClient();
+		DB database = mongoClient.getDB("studentsDB");
+		DBCollection collection = database.getCollection("studentsCollection");
 
 		if (args.length != 0) {
-			// If the passed path is incorrect and there is available bkp file
-			if (!(fileExists(args[0])) && (bkpFile.exists() && !(bkpFile.isDirectory()))) {
-				printFromBkp(bkpFile);
+			// If the passed path is incorrect and the db is not empty
+			if (!(fileExists(args[0])) && (collection.count() != 0)) {
+				printFromBkp(collection);
 			}
 		}
 
@@ -31,7 +35,7 @@ public class Main {
 			// If the file exists parsing the JSON in StudentGroup
 			StudentGroup students = gson.fromJson(br, StudentGroup.class);
 			// Creating bkp file with the available information
-			createBackup(students, bkpFile);
+			createBackup(students, collection);
 
 			if (args.length == 1) {
 				// If only one argument is passed printing the entire info
@@ -73,27 +77,17 @@ public class Main {
 		}
 	}
 
-	private static void printFromBkp(File bkpFile) {
-		Gson gson = new Gson();
-		try {
-			BufferedReader br = new BufferedReader(new FileReader(bkpFile));
-			StudentGroup studentsBkp = gson.fromJson(br, StudentGroup.class);
-			printAllElements(studentsBkp);
-		} catch (FileNotFoundException e) {
-			e.printStackTrace();
+	private static void printFromBkp(DBCollection collection) {
+		DBCursor cursor = collection.find();
+		while (cursor.hasNext()) {
+			// should call a method for reverting from DB obj to Person obj - will check
+			// Morphia
 		}
 	}
 
-	private static void createBackup(StudentGroup students, File bkpFile) {
-		try {
-			Gson gson = new Gson();
-			String jsonString = gson.toJson(students);
-			FileWriter fileWriter = new FileWriter(bkpFile);
-			fileWriter.write(jsonString);
-			fileWriter.close();
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
+	private static void createBackup(StudentGroup students, DBCollection collection) {
+		// should clear the DB before creating a new bkp
+		students.getStudents().forEach(student -> collection.insert(StudentAdaptor.toDBObject(student)));
 	}
 
 	public static void printAllElements(StudentGroup students) {
