+++ src/main/java/com/musala/simple/students/db/StudentGroup.java
@@ -1,4 +1,4 @@
-package com.musala.simple.students;
+package com.musala.simple.students.db;
 
 import java.util.List;
